move sendRequestCoro into source file
