+++ lib/inc/drogon/HttpClient.h
@@ -34,27 +34,6 @@ namespace drogon
 class HttpClient;
 using HttpClientPtr = std::shared_ptr<HttpClient>;
 
-namespace internal
-{
-#ifdef __cpp_impl_coroutine
-struct HttpRespAwaiter : public CallbackAwaiter<HttpResponsePtr>
-{
-    HttpRespAwaiter(HttpClient *client, HttpRequestPtr req, double timeout)
-        : client_(client), req_(std::move(req)), timeout_(timeout)
-    {
-    }
-
-    void await_suspend(std::coroutine_handle<> handle);
-
-  private:
-    HttpClient *client_;
-    HttpRequestPtr req_;
-    double timeout_;
-};
-#endif
-
-}  // namespace internal
-
 /// Asynchronous http client
 /**
  * HttpClient implementation object uses the HttpAppFramework's event loop by
@@ -154,12 +133,7 @@ class HttpClient : public trantor::NonCopyable
      * @return task<HttpResponsePtr>
      */
     Task<HttpResponsePtr> sendRequestCoro(HttpRequestPtr req,
-                                          double timeout = 0)
-    {
-        co_return co_await internal::HttpRespAwaiter(this,
-                                                     std::move(req),
-                                                     timeout);
-    }
+                                          double timeout = 0);
 #endif
 
     /// Set the pipelining depth, which is the number of requests that are not
@@ -268,37 +242,6 @@ class HttpClient : public trantor::NonCopyable
 };
 
 #ifdef __cpp_impl_coroutine
-inline void internal::HttpRespAwaiter::await_suspend(
-    std::coroutine_handle<> handle)
-{
-    client_->sendRequest(
-        req_,
-        [handle = std::move(handle), this](ReqResult result,
-                                           const HttpResponsePtr &resp) {
-            if (result == ReqResult::Ok)
-            {
-                setValue(resp);
-                handle.resume();
-            }
-            else
-            {
-                std::string reason;
-                if (result == ReqResult::BadResponse)
-                    reason = "BadResponse";
-                else if (result == ReqResult::NetworkFailure)
-                    reason = "NetworkFailure";
-                else if (result == ReqResult::BadServerAddress)
-                    reason = "BadServerAddress";
-                else if (result == ReqResult::Timeout)
-                    ;
-                reason = "Timeout";
-                setException(
-                    std::make_exception_ptr(std::runtime_error(reason)));
-                handle.resume();
-            }
-        },
-        timeout_);
-}
 #endif
 
 }  // namespace drogon
